pyproject.toml
[project]
name = "mcpserver"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.28.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.0",
]
